% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{cell.level_boxplots}
\alias{cell.level_boxplots}
\title{Data visualization: individual morphology measures}
\usage{
cell.level_boxplots(data, group)
}
\arguments{
\item{data}{is your input data frame}

\item{group}{is the variable you want to group your data by}
}
\description{
Linear mixed model to statistically assess how your experimental variables of interest
influence each morphology measure, at the cell level
}
