% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{celldensity}
\alias{celldensity}
\title{Calculate microglia density}
\usage{
celldensity(AreasPath, SamplesizeDF)
}
\arguments{
\item{AreasPath}{is the path to your Areas.csv file output from MicrogliaMorphology}

\item{SampleSizeDF}{is the dataframe of cell numbers that you generated from the 'samplesize' function within MicrogliaMorphologyR}
}
\description{
'celldensity' loads in your Areas.csv file from MicrogliaMorphology, which contains the areas of your images or
rois within images, then calculates the density of microglia cells for each image
}
