% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{clusterfeatures}
\alias{clusterfeatures}
\title{Cluster-specific average morphology measures}
\usage{
clusterfeatures(data, start, end)
}
\arguments{
\item{data}{is your input data frame}

\item{start}{is first column number of morphology measures}

\item{end}{is last column number of morphology measures}
}
\description{
'clusterfeatures' generates heatmap visualization of average cluster-specific morphology measures relative to other clusters
}
