% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{clusterpercentage_boxplots}
\alias{clusterpercentage_boxplots}
\title{How do morphology cluster percentages vary across variables of interest?}
\usage{
clusterpercentage_boxplots(data, ...)
}
\arguments{
\item{data}{is your input data frame}

\item{...}{list out your variables of interest without combining (e.g., Cohort, Sex, Treatment)}
}
\description{
'clusterpercentage_boxplots' generates boxplots that depict cluster shifts with experimental variables of interest
}
