% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{clusterplots}
\alias{clusterplots}
\title{Plot to show morphology clusters in PC space}
\usage{
clusterplots(data, pc.xaxis, pc.yaxis)
}
\arguments{
\item{data}{is your input data frame}

\item{pc.xaxis}{is the pc values you want on x-axis (e.g, PC1)}

\item{pc.yaxis}{is the pce values you want on y-axis (e.g., PC2)}
}
\description{
'clusterplots' visualizes k-means morphology clusters in PC space
}
