% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2xLPS}
\alias{data_2xLPS}
\title{2xLPS mouse dataset from frontal cortex, striatum, and hippocampal subregions}
\format{
\subsection{\code{data_2xLPS}}{

A data frame with 46,104 rows and 35 columns:
\describe{
\item{Antibody}{Microglia marker used in immunofluorescent image}
\item{MouseID}{Subject identifier}
\item{Sex}{Sex}
\item{Treatment}{PBS vehicle or 2xLPS i.p. injection at 0.5 mg/kg, spaced 24 hours apart}
\item{BrainRegion}{frontal cortex, striatum, or hippocampus}
\item{Subregion}{frontal cortex: infralimbic (IL), prelimbic (PL), anterior cingulate cortex (ACC); hippocampus: CA1, CA2, CA3, dentate gyrus (DG); stratium: caudate putamen (CP), nucleus accumbens (NA)}
...
}
}
}
\usage{
data_2xLPS
}
\description{
2xLPS mouse dataset from frontal cortex, striatum, and hippocampal subregions
}
\keyword{datasets}
