% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_2xLPS_fuzzykmeans}
\alias{data_2xLPS_fuzzykmeans}
\title{Fkm dataset from 2xLPS tided up}
\format{
\subsection{\code{data_2xLPS_fuzzykmeans}}{

A data frame with 46,104 rows and 35 columns:
\describe{
\item{Antibody}{Microglia marker used in immunofluorescent image}
\item{MouseID}{Subject identifier}
\item{BrainRegion}{frontal cortex, striatum, or hippocampus}
\item{Subregion}{frontal cortex: infralimbic (IL), prelimbic (PL), anterior cingulate cortex (ACC); hippocampus: CA1, CA2, CA3, dentate gyrus (DG); stratium: caudate putamen (CP), nucleus accumbens (NA)}
...
}
}
}
\usage{
data_2xLPS_fuzzykmeans
}
\description{
Fkm dataset from 2xLPS tided up
}
\keyword{datasets}
