% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_EAE}
\alias{data_EAE}
\title{EAE mouse dataset from frontal cortex, striatum, and hippocampal subregions}
\format{
\subsection{\code{data_EAE}}{

A data frame with 46,104 rows and 35 columns:
\describe{
\item{Antibody}{Microglia marker used in immunofluorescent image}
\item{MouseID}{Subject identifier}
\item{BrainRegion}{frontal cortex, striatum, or hippocampus}
\item{Subregion}{frontal cortex: infralimbic (IL), prelimbic (PL), anterior cingulate cortex (ACC); hippocampus: CA1, CA2, CA3, dentate gyrus (DG); stratium: caudate putamen (CP), nucleus accumbens (NA)}
...
}
}
}
\usage{
data_EAE
}
\description{
EAE mouse dataset from frontal cortex, striatum, and hippocampal subregions
}
\keyword{datasets}
