% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{featurecorrelations_stats}
\alias{featurecorrelations_stats}
\title{Underlying stats for correlation heatmap across morphology features}
\usage{
featurecorrelations_stats(data, featurestart, featureend, rthresh, pthresh)
}
\arguments{
\item{data}{is your input data frame}

\item{featurestart}{is first column number of morphology measures}

\item{featureend}{is last column number of morphology measures}

\item{rthresh}{is cutoff threshold for significant correlation values}

\item{pthresh}{is cutoff threshold for significant p-values}
}
\description{
This function is complementary to the 'featurecorrelations' function. 'featurecorrelations_stats'
outputs a dataframe containing the correlation and pvalues related to your featurecorrelations heatmap.
}
