% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{fraclac_tidying}
\alias{fraclac_tidying}
\title{Tidy up your FracLac output}
\usage{
fraclac_tidying(dir)
}
\arguments{
\item{dir}{is your directory containing the 'Hull and Circles Results.txt' file from FracLac output}
}
\description{
'fraclac_tidying' loads in your FracLac output and cleans up names of cells
}
