% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{logscale}
\alias{logscale}
\title{Log scale}
\usage{
logscale(x, y)
}
\arguments{
\item{x}{is your input values}

\item{y}{is the constant value you want to add (e.g., 0, 0.1, 1)}
}
\description{
'logscale' allows you to log scale values
}
