% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{merge_data}
\alias{merge_data}
\title{Merge FracLac and AnalyzeSkeleton results together into one final cell-level data frame}
\usage{
merge_data(fraclac, skeleton)
}
\arguments{
\item{fraclac}{is your tidied fraclac output}

\item{skeleton}{is your tidied skeleton output}
}
\description{
'merge_data' merges your FracLac and AnalyzeSkeleton results by Name and ID, gets rid of non-numerical data, and brushes up feature names.
The final output is a dataframe, where every row is a cell and every column is an identifier or one of 27 unique morphology features.
}
