% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{metadata_columns}
\alias{metadata_columns}
\title{Would you like to get some metadata?}
\usage{
metadata_columns(x, y, z)
}
\arguments{
\item{x}{is your tidied dataframe containing all of your numerical morphology data}

\item{y}{is your metadata listed out in order that they appear in 'Name' column: e.g., c("Cohort","Sex","MouseID")}

\item{z}{is the character, "_" or "-", that separates your metadata}
}
\description{
This function allows you to extract metadata columns from the UniqueID column.
}
