% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{normalize_logplots}
\alias{normalize_logplots}
\title{Exploratory data analysis: normalization methods}
\usage{
normalize_logplots(x)
}
\arguments{
\item{x}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
This function allows you to check distributions of features after log normalization (log10(x+1))
}
