% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{normalize_minmax}
\alias{normalize_minmax}
\title{Exploratory data analysis: min-max scaled plots}
\usage{
normalize_minmax(data)
}
\arguments{
\item{data}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
'normalize_minmax' allows you to check distributions of features after min-max normalizing to get your values for each measure constrained to 0-1 range.
}
