% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{normalize_scaled}
\alias{normalize_scaled}
\title{Exploratory data analysis: scaled plots}
\usage{
normalize_scaled(data)
}
\arguments{
\item{data}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
'normalize_scaled' allows you to check distributions of features after scaling (e.g., scale(value))
}
