% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{outliers_boxplots}
\alias{outliers_boxplots}
\title{Exploratory data analysis: outlier detection}
\usage{
outliers_boxplots(data)
}
\arguments{
\item{data}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
'outliers_boxplots' generates boxplots of morphology features to visualize those with broader ranges which might dominate the analysis downstream.
}
