% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{outliers_distributions}
\alias{outliers_distributions}
\title{Exploratory data analysis: outlier detection}
\usage{
outliers_distributions(x)
}
\arguments{
\item{x}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
This function allows you to check distributions of features to determine if and how to normalize
}
