% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{pcadata}
\alias{pcadata}
\title{Dimensionality reduction using PCA}
\usage{
pcadata(data, featurestart, featureend, pc.start, pc.end)
}
\arguments{
\item{data}{is your input data frame}

\item{featurestart}{is first column number of morphology measures}

\item{featureend}{is last column number of morphology measures}

\item{pc.start}{is first PC you want included (e.g., PC1)}

\item{pc.end}{is last PC you want included (e.g., PC4)}
}
\description{
'pcadata' allows you to perform PCA analysis and update your dataframe to include PCs of interest
}
