% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{pcadata_elbow}
\alias{pcadata_elbow}
\title{Dimensionality reduction using PCA and elbow/scree plot to get variance explained}
\usage{
pcadata_elbow(data, start, end)
}
\arguments{
\item{data}{is your input data frame}

\item{start}{is first column number of morphology measures}

\item{end}{is last column number of morphology measures}
}
\description{
'pcadata_elbow' allows you to perform PCA analysis and update your dataframe
}
