% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{pcfeaturecorrelations}
\alias{pcfeaturecorrelations}
\title{What morphology features describe the PCs?}
\usage{
pcfeaturecorrelations(
  pca_data,
  pc.start,
  pc.end,
  feature.start,
  feature.end,
  rthresh,
  pthresh,
  title
)
}
\arguments{
\item{pc.start}{is first PC you want included in analysis (e.g., PC1)}

\item{pc.end}{is last PC you want included in analysis (e.g., PC4)}

\item{feature.start}{is first column number of morphology measures}

\item{feature.end}{is last column number of morphology measures}

\item{rthresh}{is cutoff threshold for significant correlation values}

\item{pthresh}{is cutoff threshold for significant p-values}

\item{title}{is what you want to name your heatmap}

\item{data}{is your input data frame}
}
\description{
'pcfeaturecorrelations' allows you to perform correlations between the top PCs and morphology features to assess which features are driving the variability captured by each of the PCs.
}
