% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{pcfeaturecorrelations_stats}
\alias{pcfeaturecorrelations_stats}
\title{Underlying stats for correlation heatmap exploring which morphologies describe the PCs}
\usage{
pcfeaturecorrelations_stats(
  pca_data,
  pc.start,
  pc.end,
  feature.start,
  feature.end,
  rthresh,
  pthresh
)
}
\arguments{
\item{pc.start}{is first PC you want included in analysis (e.g., PC1)}

\item{pc.end}{is last PC you want included in analysis (e.g., PC4)}

\item{feature.start}{is first column number of morphology measures}

\item{feature.end}{is last column number of morphology measures}

\item{rthresh}{is cutoff threshold for significant correlation values}

\item{pthresh}{is cutoff threshold for significant p-values}

\item{data}{is your input data frame}
}
\description{
This function is complementary to the 'pcfeaturecorrelations' function. 'pcfeaturecorrelations_stats'
outputs a dataframe containing the correlation and pvalues related to your pcfeaturecorrelations heatmap.
}
