% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{plots_expvariable}
\alias{plots_expvariable}
\title{Explore how experimental variables describe data}
\usage{
plots_expvariable(data, pc.xaxis, pc.yaxis)
}
\arguments{
\item{data}{is your input data frame}

\item{pc.xaxis}{is the pc values you want on x-axis (e.g, PC1)}

\item{pc.yaxis}{is the pc values you want on y-axis (e.g., PC2)}
}
\description{
'plots_expvariable' colors points in pc space by experimental variables
}
