% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{samplesize}
\alias{samplesize}
\title{Sample size}
\usage{
samplesize(data, ...)
}
\arguments{
\item{data}{is your input dataframe}

\item{...}{list out your variables of interest without combining (e.g., Cohort, Sex, Treatment)}
}
\description{
'samplesize' allows you to obtain sample size of your experimental groups based on variables of interest
}
