% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{skeleton_tidying}
\alias{skeleton_tidying}
\title{Concatenate your individual AnalyzeSkeleton output files into one dataframe}
\usage{
skeleton_tidying(dir)
}
\arguments{
\item{dir}{is your directory containing all of your individual AnalyzeSkeleton .csv result files output from MicrogliaMorphology}
}
\description{
'skeleton_tidying' loads in all of your AnalyzeSkeleton output and clean up IDs to match FracLac data for downstream merging of data at cell level.
}
