% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{stats_cluster.animal}
\alias{stats_cluster.animal}
\title{Stats analysis: cluster-level changes}
\usage{
stats_cluster.animal(data, model, posthoc.sex, posthoc.nosex, adjust)
}
\arguments{
\item{data}{is your input data frame}

\item{model}{is your linear mixed model (e.g., Value~ Cluster\emph{Treatment}Sex + (1|MouseID))}

\item{posthoc.sex}{is your posthoc comparisons when considering sex (e.g., ~Cluster*Treatment|Sex)}

\item{posthoc.nosex}{is your posthoc comparisons when not considering sex (e.g., ~Cluster*Treatment)}

\item{adjust}{is your method of multiple test correction}
}
\description{
Linear mixed model to statistically assess how your experimental variables of interest
influence cluster percentages, at animal level
}
