% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{stats_morphologymeasures.cell}
\alias{stats_morphologymeasures.cell}
\title{Stats analysis: individual morphology measures}
\usage{
stats_morphologymeasures.cell(data, model, posthoc1, posthoc2, adjust)
}
\arguments{
\item{data}{is your input data frame}

\item{model}{is your linear mixed model (e.g., Value ~ Treatment*Sex + (1|MouseID))}

\item{posthoc1}{is your posthoc comparisons when considering sex (e.g., ~Treatment|Sex)}

\item{posthoc2}{is your posthoc comparisons when not considering sex (e.g., ~Treatment)}

\item{adjust}{is your method of multiple test correction}
}
\description{
Linear mixed model to statistically assess how your experimental variables of interest
influence each morphology measure, at the cell level
}
