% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{transform_log}
\alias{transform_log}
\title{Transform data: Log scale}
\usage{
transform_log(data, x, start, end)
}
\arguments{
\item{data}{is your input dataframe}

\item{x}{is the constant value you want to add (e.g., 0.1, 1)}

\item{start}{is first column number of morphology measures}

\item{end}{is last column number of morphology measures}
}
\description{
'transform_log' allows you to log scale your dataframe
}
