% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{transform_minmax}
\alias{transform_minmax}
\title{Transform data: Min-max scaling}
\usage{
transform_minmax(data, start, end)
}
\arguments{
\item{data}{is your input dataframe}

\item{start}{is first column number of morphology measures}

\item{end}{is last column number of morphology measures}
}
\description{
'transform_minmax' allows you to min-max scale your dataframe within each morphology measure
}
