% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ImageTypeComparison}
\alias{data_ImageTypeComparison}
\title{ImageTypeComparisons mouse dataset}
\format{
\subsection{\code{data_ImageTypeComparison}}{

A data frame with 60 rows and 13 columns:
\describe{
\item{CellID}{ameboid1-5, hypertrophic1-5, ramified1-5, rod1-5}
\item{MorphologyClass}{ameboid, hypertrophic, ramified, rod}
\item{ID}{1, 2, 3, 4, 5}
\item{ImageType}{2D, 3D, EDF}
\item{AnalyzeSkeleton measures: Columns 5-13}{# of branches, # of junctions, # of end point voxels, # of junction voxels, # of slab voxels, Average branch length # of triple points, # of quadruple points, Maximum branch length}
...
}
}
}
\usage{
data_ImageTypeComparison
}
\description{
Data used for image type comparisons in Figure 2 of publication for MicrogliaMorphologyR package (Kim et al., 2024)
}
\keyword{datasets}
