% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{fraclac_tidying}
\alias{fraclac_tidying}
\title{Tidying up your FracLac output}
\usage{
fraclac_tidying(x)
}
\arguments{
\item{x}{is your directory containing the 'Hull and Circles Results.txt' file}

\item{y}{is the variable name for your processed output containing your tidied fraclac data}
}
\description{
This function allows you to load in FracLac output and clean up names of cells
}
