% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{merge_data}
\alias{merge_data}
\title{Merge fraclac and skeleton results together}
\usage{
merge_data(x, y)
}
\arguments{
\item{x}{is your tidied fraclac output}

\item{y}{is your tidied skeleton output}
}
\description{
This function allows you to merge your fraclac and skeleton results by Name and ID, get rid of non-numerical data, and brush up feature names
}
