% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR-package.R
\docType{package}
\name{MicrogliaMorphologyR-package}
\alias{MicrogliaMorphologyR}
\alias{MicrogliaMorphologyR-package}
\title{MicrogliaMorphologyR: R Package For Microglia Morphology Analysis}
\description{
MicrogliaMorphologyR is complimentary to ImageJ macro MicrogliaMorphology. You can use MicrogliaMorphologyR to perform exploratory data analysis and visualization of 27 different morphology features, characterize morphological cluster identities, quantify shifts in morphological populations, generate heatmap and boxplot visualizations of data in flexible ways including at the single-cell level, animal-level, and experimental condition-level, and perform statistical analysis of your data.
}
\author{
\strong{Maintainer}: Jennifer Kim \email{jenn_kim@outlook.com} (\href{https://orcid.org/0000-0003-2922-1010}{ORCID})

}
\keyword{internal}
