% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{normalize_logplots}
\alias{normalize_logplots}
\title{Exploratory data analysis: normalization methods}
\usage{
normalize_logplots(data, x)
}
\arguments{
\item{data}{is your final dataframe which contains gathered data (measure, value) format}

\item{x}{is the constant value you want to add (e.g., 0, 0.1, 1)}
}
\description{
'normalize_logplots' allows you to check distributions of features after log normalization (e.g., log10(value+0.1))
}
