% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{outliers_boxplots}
\alias{outliers_boxplots}
\title{Exploratory data analysis: outlier detection}
\usage{
outliers_boxplots(x)
}
\arguments{
\item{x}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
This function allows you to use boxplots of features to look for those with broader ranges which might dominate the data analysis
}
