% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{outliers_distributions}
\alias{outliers_distributions}
\title{Exploratory data analysis: outlier detection}
\usage{
outliers_distributions(data)
}
\arguments{
\item{data}{is your final dataframe which contains gathered data (measure, value) format}
}
\description{
'outliers_distributions' generates plots of feature distributions to determine which features have skewed distributions and how to normalize your values for analysis downstream.
}
