% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{skeleton_tidying}
\alias{skeleton_tidying}
\title{Concatenating your individual AnalyzeSkeleton output files into one dataframe}
\usage{
skeleton_tidying(x)
}
\arguments{
\item{x}{is your directory containing all of your individual AnalyzeSkeleton .csv result files}

\item{y}{is the variable name for your processed output containing your tidied analyzeskeleton data}
}
\description{
This function allows you to load in all of your AnalyzeSkeleton output and clean up IDs to match FracLac data for downstream merging
}
