% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{stats_cluster.animal}
\alias{stats_cluster.animal}
\title{Stats analysis: cluster-level changes}
\usage{
stats_cluster.animal(data, model, groupby, posthoc1, posthoc2, adjust)
}
\arguments{
\item{data}{is your input data frame}

\item{model}{is your linear mixed model (e.g., Value ~ Cluster\emph{Treatment}Sex + (1|MouseID))}

\item{groupby}{do you want to group by clusters so that your posthocs are run within each cluster rather than within and across?}

\item{posthoc1}{is your first set of posthoc comparisons (e.g., considering sex: ~Cluster*Treatment|Sex)}

\item{posthoc2}{is your second set of posthoc comparisons (e.g., not considering sex: ~Cluster*Treatment)}

\item{adjust}{is your method of multiple test correction}
}
\description{
Linear mixed model to statistically assess how your experimental variables of interest
influence cluster percentages, at animal level
}
