% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrogliaMorphologyR_functions.R
\name{transform_scale}
\alias{transform_scale}
\title{Transform data: Scaling}
\usage{
transform_scale(data, start, end)
}
\arguments{
\item{data}{is your input dataframe}

\item{start}{is first column number of morphology measures}

\item{end}{is last column number of morphology measures}
}
\description{
'transform_scale' allows you to scale your dataframe within each morphology measure
}
